requirements.txt
ujson
tensorflow-gpu
keras
profilehooks
numpy
python-chess
h5py